Rename ColourPalette.h to PaletteType.h
