+++ src/openrct2/drawing/BlendColourMap.cpp
@@ -14,8 +14,8 @@
 #ifndef DISABLE_TTF
 
     #include "../core/EnumUtils.hpp"
-    #include "ColourPalette.h"
     #include "Drawing.h"
+    #include "PaletteType.h"
 
     #include <cmath>
