+++ src/openrct2/drawing/Drawing.h
@@ -11,9 +11,9 @@
 
 #include "../core/CallingConventions.h"
 #include "Colour.h"
-#include "ColourPalette.h"
 #include "ImageId.hpp"
 #include "PaletteMap.h"
+#include "PaletteType.h"
 
 #include <optional>
 #include <span>
