+++ src/openrct2/drawing/PaletteType.h
